requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-WTF==1.2.2
WTForms==3.2.1
SQLAlchemy==2.0.36
qrcode[pil]==8.0
Pillow==11.1.0
reportlab==4.2.5
bcrypt==4.2.1
Authlib==1.3.0
requests==2.31.0
cryptography==44.0.0
itsdangerous==2.2.0
email-validator==2.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
cloudinary==1.41.0
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
